+++ ngblocks/get.go
@@ -5,6 +5,8 @@ import (
 	"fmt"
 
 	"github.com/dgraph-io/badger/v2"
+	"google.golang.org/protobuf/proto"
+
 	"github.com/ngchain/ngcore/ngtypes"
 	"github.com/ngchain/ngcore/utils"
 )
@@ -119,4 +121,87 @@ func GetLatestHash(txn *badger.Txn) ([]byte, error) {
 	}
 
 	return hash, nil
+}
+
+func GetLatestBlock(txn *badger.Txn) (*ngtypes.Block, error) {
+	hash, err := GetLatestHash(txn)
+	if err != nil {
+		return nil, err
+	}
+
+	block, err := GetBlockByHash(txn, hash)
+	if err != nil {
+		return nil, err
+	}
+
+	return block, nil
+}
+
+func GetOriginHeight(txn *badger.Txn) (uint64, error) {
+	key := append(blockPrefix, originHeightTag...)
+	item, err := txn.Get(key)
+	if err == badger.ErrKeyNotFound {
+		return 0, err // export the keynotfound
+	}
+	if err != nil {
+		return 0, fmt.Errorf("failed to get latest height: %s", err)
+	}
+	raw, err := item.ValueCopy(nil)
+	if err != nil {
+		return 0, fmt.Errorf("no such height in latestTag: %s", err)
+	}
+
+	return binary.LittleEndian.Uint64(raw), nil
+}
+
+func GetOriginHash(txn *badger.Txn) ([]byte, error) {
+	key := append(blockPrefix, originHashTag...)
+	item, err := txn.Get(key)
+	if err == badger.ErrKeyNotFound {
+		return nil, err // export the keynotfound
+	}
+	if err != nil {
+		return nil, fmt.Errorf("failed to get latest hash: %s", err)
+	}
+	hash, err := item.ValueCopy(nil)
+	if err != nil {
+		return nil, fmt.Errorf("failed to get latest hash: %s", err)
+	}
+
+	return hash, nil
+}
+
+func GetOriginBlock(txn *badger.Txn) (*ngtypes.Block, error) {
+	key := append(blockPrefix, originHashTag...)
+	item, err := txn.Get(key)
+	if err == badger.ErrKeyNotFound {
+		return nil, err // export the keynotfound
+	}
+	if err != nil {
+		return nil, fmt.Errorf("failed to get origin block hash: %s", err)
+	}
+	hash, err := item.ValueCopy(nil)
+	if err != nil {
+		return nil, fmt.Errorf("failed to get origin block hash: %s", err)
+	}
+
+	item, err = txn.Get(append(blockPrefix, hash...))
+	if err == badger.ErrKeyNotFound {
+		return nil, err // export the keynotfound
+	}
+	if err != nil {
+		return nil, fmt.Errorf("failed to get origin block hash: %s", err)
+	}
+	rawBlock, err := item.ValueCopy(nil)
+	if err != nil {
+		return nil, fmt.Errorf("failed to get origin block: %s", err)
+	}
+
+	var block ngtypes.Block
+	err = proto.Unmarshal(rawBlock, &block)
+	if err != nil {
+		return nil, fmt.Errorf("failed to get origin block: %s", err)
+	}
+
+	return &block, nil
 }