+++ ngchain/check_health.go
@@ -3,14 +3,21 @@ package ngchain
 import (
 	"bytes"
 	"fmt"
+
 	"github.com/ngchain/ngcore/ngtypes"
 )
 
 func (chain *Chain) CheckHealth(network ngtypes.NetworkType) {
 	log.Warn("checking chain's health")
 	latestHeight := chain.GetLatestBlockHeight()
-	prevBlockHash := ngtypes.GetGenesisBlock(network).PrevBlockHash
-	for h := uint64(0); h <= latestHeight; h++ {
+
+	origin := chain.GetOriginBlock()
+	originHeight := origin.Height
+
+	prevBlockHash := origin.Hash()
+
+	for h := originHeight; h < latestHeight; {
+		h++
 		b, err := chain.GetBlockByHeight(h)
 		if err != nil {
 			panic(err)