+++ ngtypes/defaults.go
@@ -21,9 +21,12 @@ var (
 // PoW const
 const (
 	// MinimumDifficulty is the minimum of pow minimumDifficulty because my laptop has 200 h/s, I believe you can either
-	minimumDifficulty = 200 << 4         // Target = MaxTarget / diff
-	TargetTime        = 16 * time.Second // change time from 10 -> 16 = 1 << 4
-	BlockCheckRound   = 10               // do fork if fall behind one round
+	minimumDifficulty = 200 << 4 // Target = MaxTarget / diff
+	TargetTime        = 16 * time.Second
+	BlockCheckRound   = 10 // do converge if fall behind one round
+
+	MatureRound  = 10                            // not mandatory required, can be modified by different daemons
+	MatureHeight = MatureRound * BlockCheckRound // just for calculating the immature balance
 )
 
 // PoW variables
@@ -36,10 +39,9 @@ var (
 
 // Maximum sizes
 const (
-	// !NO MAX LIMITATION!
-	//BlockMaxTxsSize = 1 << 25 // 32M
+	// BlockMaxTxsSize = 1 << 25 // 32M
 	TxMaxExtraSize = 1 << 20 // if more than 1m, extra should be separated ot multi append
-	// The length of a timestemp bytes
+	// The length of a timestamp bytes
 	TimestampSize = 8
 	// The length of a hash bytes
 	HashSize = 32
@@ -90,10 +92,10 @@ func GetGenesisGenerateTxSignature(network NetworkType) []byte {
 func GetGenesisBlockNonce(network NetworkType) []byte {
 	switch network {
 	case NetworkType_ZERONET:
-		genesisBlockNonce, _ := hex.DecodeString("c800120f3ae9a2fc")
+		genesisBlockNonce, _ := hex.DecodeString("84e1be18c794f125")
 		return genesisBlockNonce
 	case NetworkType_TESTNET:
-		genesisBlockNonce, _ := hex.DecodeString("115c488d6d09dc41")
+		genesisBlockNonce, _ := hex.DecodeString("55e0414311982f0e")
 		return genesisBlockNonce
 	case NetworkType_MAINNET:
 		panic("not ready for mainnet")
@@ -115,4 +117,13 @@ func GetGenesisTimestamp(network NetworkType) int64 {
 	default:
 		panic("unknown network")
 	}
+}
+
+// 100 * X
+func GetMatureHeight(currentHeight uint64) uint64 {
+	if currentHeight < MatureHeight {
+		return 0
+	}
+
+	return currentHeight / MatureHeight * MatureHeight
 }